🚚 게임 아이디 연동 관련 type 분리
