+++ src/components/find/main-card/FindCard.tsx
@@ -23,7 +23,7 @@ import FindDetailModal from "./FindDetailModal";
 import { CreateChat } from "@/services/chat.client";
 import { useQuery } from "@tanstack/react-query";
 import { getPartyDetail } from "@/services/party.client";
-import { GameAccount } from "@/types/user";
+import { GameAccount } from "@/types/game-account";
 import { useRouter } from "next/navigation";
 
 interface FindCardProps extends HTMLAttributes<HTMLDivElement> {
