+++ src/types/game-account.ts
@@ -1,0 +1,12 @@
+export type GameAccount = {
+  gameAccountId: number;
+  gameNickname: string;
+  gameTag: string;
+  gameType: string;
+  puuid: string;
+  profileIconId: number;
+  profileIconUrl: string;
+  userId: number;
+  createdAt: string;
+  updatedAt: string;
+};
