+++ src/types/user.ts
@@ -22,16 +22,3 @@ export type MiniProfile = {
   };
   isBlockedByMe: boolean;
 };
-
-export type GameAccount = {
-  gameAccountId: number;
-  gameNickname: string;
-  gameTag: string;
-  gameType: string;
-  puuid: string;
-  profileIconId: number;
-  profileIconUrl: string;
-  userId: number;
-  createdAt: string;
-  updatedAt: string;
-};
